app/requirements.txt
flask
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors
flask-login
SQLAlchemy
mysqlclient
cryptography
werkzeug
gunicorn
flask-swagger-ui==3.36.0
python-dotenv